some postgres

+++ lesson4/internal/pkg/item/usecase/delete.go
@@ -5,9 +5,8 @@ import (
 	"fmt"
 )
 
-func (u usecase) Delete(ctx context.Context, id int) error {
+func (u usecase) Delete(ctx context.Context, id int64) error {
 	defer u.stat.MethodDuration.WithLabels(map[string]string{"method_name": "Delete"}).Start().Stop()
-
 	if err := u.repo.Delete(ctx, id); err != nil {
 		return fmt.Errorf("failed to delete item in repo: %w", err)
 	}
